Renames DepthwiseConv1D definition script.
